Migrate PaymentModeAttributeTypeResource to use new PaymentModeAttributeTypeService
